Move project forms/modals to separate folders
