+++ portal/src/routes/(authenticated)/projects/[id=idNumber]/forms/Reviewers.svelte
@@ -1,0 +1,90 @@
+<script lang="ts">
+  import { enhance as svk_enhance } from '$app/forms';
+  import IconContainer from '$lib/components/IconContainer.svelte';
+  import { m } from '$lib/paraglide/messages';
+  import { getLocale, locales } from '$lib/paraglide/runtime';
+  import { byName } from '$lib/utils/sorting';
+  import { superForm, type Infer, type SuperValidated } from 'sveltekit-superforms';
+  import type { ReviewerSchema } from './valibot';
+
+  interface Props {
+    reviewers: {
+      Id: number;
+      Name: string | null;
+      Email: string | null;
+    }[];
+    formData: SuperValidated<Infer<ReviewerSchema>>;
+  }
+
+  let { reviewers, formData }: Props = $props();
+
+  const { form, enhance } = superForm(formData, {
+    resetForm: true
+  });
+</script>
+
+<div class="card card-bordered border-slate-400 overflow-hidden rounded-md max-w-full">
+  <div class="bg-neutral">
+    <h2>{m.project_side_reviewers_title()}</h2>
+  </div>
+  <div class="p-2">
+    {#if reviewers.length}
+      {@const locale = getLocale()}
+      {#each reviewers.toSorted((a, b) => byName(a, b, locale)) as reviewer}
+        <div class="flex flex-row w-full place-content-between p-2">
+          <span>{reviewer.Name} ({reviewer.Email})</span>
+          <form action="?/deleteReviewer" method="post" use:svk_enhance>
+            <input type="hidden" name="id" value={reviewer.Id} />
+            <button type="submit" class="cursor-pointer">
+              <IconContainer icon="mdi:close" width="24" />
+            </button>
+          </form>
+        </div>
+      {/each}
+    {:else}
+      <p class="p-2">{m.project_side_reviewers_empty()}</p>
+    {/if}
+  </div>
+  <div class="p-2 bg-neutral">
+    <form action="?/addReviewer" method="post" use:enhance>
+      <div class="flex flex-col place-content-between space-y-2">
+        <div class="flex flex-col gap-2 reviewerform">
+          <input
+            type="text"
+            name="name"
+            placeholder="Name"
+            class="input input-bordered grow validator"
+            bind:value={$form.name}
+            required
+          />
+          <input
+            type="email"
+            name="email"
+            placeholder="Email"
+            class="input input-bordered grow validator"
+            bind:value={$form.email}
+            required
+          />
+        </div>
+        <div class="flex flex-row space-x-2">
+          <select name="locale" class="grow select select-bordered" bind:value={$form.language}>
+            {#each locales as locale}
+              <option value={locale}>{locale.split('-')[0]}</option>
+            {/each}
+          </select>
+          <button type="submit" class="btn btn-primary">
+            {m.project_side_reviewers_form_submit()}
+          </button>
+        </div>
+      </div>
+    </form>
+  </div>
+</div>
+
+<style>
+  @container (width > 450px) {
+    .reviewerform {
+      flex-direction: row;
+    }
+  }
+</style>
