+++ portal/src/routes/(authenticated)/projects/[id=idNumber]/forms/index.ts
@@ -1,0 +1,6 @@
+import Authors from './Authors.svelte';
+import OwnerGroup from './OwnerGroup.svelte';
+import Reviewers from './Reviewers.svelte';
+import Settings from './Settings.svelte';
+
+export { Authors, OwnerGroup, Reviewers, Settings };
