+++ portal/src/routes/(authenticated)/projects/[id=idNumber]/modals/DeleteProduct.svelte
