+++ portal/src/routes/(authenticated)/projects/[id=idNumber]/modals/Properties.svelte
