+++ portal/src/routes/(authenticated)/projects/[id=idNumber]/modals/index.ts
@@ -1,0 +1,5 @@
+import AddProduct from './AddProduct.svelte';
+import DeleteProduct from './DeleteProduct.svelte';
+import Properties from './Properties.svelte';
+
+export { AddProduct, DeleteProduct, Properties };
